added csep1 ntest example from CSEP1 codebase updated california_relm_region to allow for user specified grid spacing renamed time module to time_utils to avoid collisions with default module added short-circuit return values for empty catalogs fixed obvious issues with JMA catalog, but still needs testing
