+++ csep/core/forecasts.py
@@ -1,6 +1,12 @@
-import numpy
-from csep.utils.log import LoggingMixin
+import itertools
 
+import numpy
+import copy
+import datetime
+from csep.utils.log import LoggingMixin
+from csep.utils.spatial import CartesianGrid2D
+from csep.utils.basic_types import Polygon
+from csep.utils.time_utils import decimal_year
 
 class GriddedDataSet(LoggingMixin):
     """Represents space-magnitude discretized seismicity implementation.
@@ -20,18 +26,18 @@ class GriddedDataSet(LoggingMixin):
               magnitudes. The magnitude bins should be regularly spaced.
     """
 
-    def __init__(self, data=None, region=None, name=None, time_horizon=None):
+    def __init__(self, data=None, region=None, name=None):
         """ Constructs GriddedSeismicity class.
 
         Attributes:
             data (numpy.ndarray): numpy.ndarray
-            grids (tuple): tuple of
+            region:
+            name:
+            time_horizon:
         """
+        super().__init__()
         self._data = data
-        # metadata associated with spatial region
         self.region = region
-        # discretization of the magnitude bins
-        self.time_horizon = time_horizon
         self.name = name
 
     @property
@@ -56,7 +62,12 @@ class GriddedDataSet(LoggingMixin):
     def longitudes(self):
         return self.region.origins()[:,0]
 
-    def scale(self, val):
+    @property
+    def polygons(self):
+        return self.region.polygons
+
+
+    def scale(self, val, in_place=False):
         """Scales forecast by floating point value.
 
         Args:
@@ -67,7 +78,12 @@ class GriddedDataSet(LoggingMixin):
         """
         if not isinstance(val, (int, float, numpy.ndarray)):
             raise ValueError("scaling value must be (int, float, or numpy.ndarray).")
-        self._data *= val
+        if not in_place:
+            new = copy.deepcopy(self)
+            new._data *= val
+            return new
+        else:
+            self._data *= val
         return self
 
     def to_dict(self):
@@ -85,11 +101,18 @@ class MarkedGriddedDataSet(GriddedDataSet):
 
     """
 
-    def __init__(self, magnitudes=None, name=None, time_horizon=None, *args, **kwargs):
+    def __init__(self, magnitudes=None, time_horizon=None, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.name = name
         self.time_horizon = time_horizon
         self.magnitudes = magnitudes
+
+    @property
+    def num_mag_bins(self):
+        return len(self.magnitudes)
+
+    @property
+    def num_nodes(self):
+        return self.region.num_nodes
 
     def spatial_counts(self, cartesian=False):
         """
@@ -103,23 +126,99 @@ class MarkedGriddedDataSet(GriddedDataSet):
 
         """
         if cartesian:
-            return self.region.get_cartesian(self.spatial_counts, cartesian=False)
+            return self.region.get_cartesian(numpy.sum(self.data, axis=1))
         else:
             return numpy.sum(self.data, axis=1)
 
     def magnitude_counts(self):
         return numpy.sum(self.data, axis=0)
 
+
+class GriddedForecast(MarkedGriddedDataSet):
+
+    def __init__(self, start_time, end_time, *args, **kwargs):
+        """
+        Constructor for GriddedForecast class
+
+        Args:
+            start_time (datetime.datetime):
+            end_time (datetime.datetime):
+        """
+        super().__init__(*args, **kwargs)
+        self.start_time = start_time
+        self.end_time = end_time
+
+    def scale_to_test_date(self, test_datetime, in_place=True):
+        """ Scales forecast data by the fraction of the date.
+
+        Uses the concept of decimal years to keep track of leap years. See the csep.utils.time_utils.decimal_year for
+        details on the implementation. If datetime is before the start_date or after the end_date, we will scale the
+        forecast by unity.
+
+        These datetime objects can be timezone aware in UTC timezone or both not time aware. This function will raise a
+        TypeError according to the specifications of datetime module if these conditions are not met.
+
+        Args:
+            test_datetime (datetime.datetime): date to scale the forecast
+            in_place (bool): if false, creates a deep copy of the object and scales that instead
+        """
+
+        # Note: this will throw a TypeError if datetimes are not either both time aware or not time aware.
+        if test_datetime >= self.end_time:
+            return self
+
+        if test_datetime <= self.start_time:
+            return self
+
+        fore_dur = decimal_year(self.end_time) - decimal_year(self.start_time)
+        # we are adding one day, bc tests are considered to occur at the end of the day specified by test_datetime.
+        test_date_dec = decimal_year(test_datetime + datetime.timedelta(1))
+        fore_frac = (test_date_dec - decimal_year(self.start_time)) / fore_dur
+        res = self.scale(fore_frac, in_place=in_place)
+        return res
+
     @classmethod
-    def from_custom(cls, afunc, name=None, time_horizon=None, *args, **kwargs):
+    def from_custom(cls, afunc, start_date, end_date, name=None, time_horizon=None, *args, **kwargs):
         """Creates MarkedGriddedDataSet class from custom parsing function.
 
         Custom parsing function should return a tuple containing the forecast data as appropriate numpy.ndarray and
         region class. We can only rely on some heuristics to ensure that these classes are set up appropriately.  """
         data, region, magnitudes = afunc(*args, **kwargs)
         # try to ensure that data are region are compatible with one another, but we can only rely on heuristics
-        return cls(data=data, region=region, magnitudes=magnitudes, name=name, time_horizon=time_horizon)
+        return cls(start_date, end_date, data=data, region=region, magnitudes=magnitudes, name=name, time_horizon=time_horizon)
 
     @classmethod
-    def from_csep1_xml(cls, xml_fname):
-        raise NotImplementedError()
+    def from_csep1_ascii(cls, ascii_fname, start_date, end_date):
+        """ Reads Forecast file from CSEP1 ascii format.
+
+        The ascii format from CSEP1 testing centers. The ASCII format does not contain headers. The format is listed here:
+
+        Lon_0, Lon_1, Lat_0, Lat_1, z_0, z_1, Mag_0, Mag_1, Rate, Flag
+
+        For the purposes of defining region objects and magnitude bins use the Lat_0 and Lon_0 values along with Mag_0.
+        We can assume that the magnitude bins are regularly spaced to allow us to compute Deltas.
+
+        The file is row-ordered so that magnitude bins are fastest then followed by space.
+
+        Args:
+            ascii_fname: file name of csep forecast in .dat format
+        """
+        # Load data
+        data = numpy.loadtxt(ascii_fname)
+        unique_poly = numpy.unique(data[:,:4], axis=0)
+        # create magnitudes bins using Mag_0, ignoring Mag_1 bc they are regular until last bin. we dont want binary search for this
+        mws = numpy.unique(data[:,-4])
+        # csep1 stores the lat lons as min values and not (x,y) tuples
+        bboxes = [tuple(itertools.product(bbox[:2], bbox[2:])) for bbox in unique_poly]
+        # the spatial cells are arranged fast in latitude, so this only works for the specific csep1 file format
+        dh = unique_poly[0,3] - unique_poly[0,2]
+        # create CarteisanGrid of points
+        region = CartesianGrid2D([Polygon(bbox) for bbox in bboxes], dh)
+        # get dims of 2d np.array
+        n_mag_bins = len(mws)
+        n_poly = region.num_nodes
+        # reshape rates into correct 2d format
+        rates = data[:,-2].reshape(n_poly,n_mag_bins)
+        # create / return class
+        gds = cls(start_date, end_date, magnitudes=mws, name=ascii_fname[:-4], region=region, data=rates)
+        return gds
