+++ csep/utils/basic_types.py
@@ -1,5 +1,6 @@
 import numpy as np
 import matplotlib
+import matplotlib.path
 
 import pyproj
