+++ requirements.yaml
@@ -16,3 +16,4 @@ dependencies:
   - cartopy=0.17.0
   - vcrpy=2.0.1
   - gitpython=3.0.5
+  - six=1.14.0
